Add "find-esp=<device>" command to find the esp partition
The <device> can be any partition or drive.  If more than one esp
partition exists on that drive, return the first one.  Some error
messages may be sent to stderr.  Use an empty string on stdout to
signal a missing esp partition.  I could modify the code to do this
but currently errors indicate something is seriously wrong while an
empty string signfies that all is well but no esp partition was found.